feat: restore one confirmation modal for all cases
