+++ components.d.ts
@@ -16,12 +16,12 @@ declare module '@vue/runtime-core' {
     BAvatar: typeof import('bootstrap-vue-next')['BAvatar']
     ConfirmationDismissalModal: typeof import('./src/components/modals/ConfirmationDismissalModal.vue')['default']
     ConfirmationModal: typeof import('./src/components/modals/ConfirmationModal.vue')['default']
-    ConfirmationWithMessageModal: typeof import('./src/components/modals/ConfirmationWithMessageModal.vue')['default']
-    ConfirmationWithMessagModal: typeof import('./src/components/modals/ConfirmationWithMessageModal.vue')['default']
+    ConfirmationWithMessageModal: typeof import('./src/components/modals/ConfirmationModal.vue')['default']
+    ConfirmationWithMessagModal: typeof import('./src/components/modals/ConfirmationModal.vue')['default']
     ConfirmationWithReasonModal: typeof import('./src/components/modals/ConfirmationWithReasonModal.vue')['default']
     copy: typeof import('./src/components/modals/ConfirmationModal copy.vue')['default']
     CopyTextButton: typeof import('./src/components/CopyTextButton.vue')['default']
-    DismissalConfirmationModal: typeof import('./src/components/modals/ConfirmationWithMessageModal.vue')['default']
+    DismissalConfirmationModal: typeof import('./src/components/modals/ConfirmationModal.vue')['default']
     FeedbackModal: typeof import('./src/components/modals/FeedbackModal.vue')['default']
     FilterSort: typeof import('./src/components/FilterSort.vue')['default']
     Footer: typeof import('./src/components/Footer.vue')['default']
